refactoring om ports
