+++ src/main/scala/diplomacy/Resources.scala
@@ -74,8 +74,6 @@ abstract class Device
   def describe(resources: ResourceBindings): Description
   /* This can be overriden to make one device relative to another */
 
-  def getOMComponents(resourceBindingsMap: ResourceBindingsMap): Seq[OMComponent] = Nil
-
   def parent: Option[Device] = None
 
   /** make sure all derived devices have an unique label */
