--- src/main/scala/diplomaticobjectmodel/DiplomaticObjectModel.scala
@@ -1,23 +1,0 @@
-// See LICENSE.SiFive for license details.
-
-package freechips.rocketchip.diplomaticobjectmodel
-
-import freechips.rocketchip.diplomaticobjectmodel.model.OMComponent
-
-import scala.collection.mutable.ListBuffer
-
-class OMCollector {
-  def getComponent(): Seq[OMComponent] = Nil
-}
-
-object DiplomaticObjectModel {
-  private val doms = ListBuffer[OMCollector]()
-
-  def add(d: OMCollector): Unit = {
-    doms += (d)
-  }
-
-  def getComponents(): OMComponent = {
-    doms.flatMap(_.getComponent()).last
-  }
-}
